+++ include/divide.h
@@ -1,5 +1,5 @@
 #pragma once
 
 namespace Calculator{
-    float divide(float a, float b);
+    double divide(double a, double b);
 }
