+++ include/multiply.h
@@ -1,5 +1,5 @@
 #pragma once
 
 namespace Calculator{
-    float multiply(float a, float b);
+    double multiply(double a, double b);
 }
